Required changes, packages refactoring, removed old files
