+++ src/main/java/it/polimi/ingsw/network/client/ClientMessageHandler.java
@@ -1,9 +1,8 @@
-package it.polimi.ingsw.network.server;
+package it.polimi.ingsw.network.client;
 
 import it.polimi.ingsw.network.client.message.ClientMessage;
 
 public interface ClientMessageHandler {
     void sendMessage(ClientMessage message);
-
     void closeConnection();
 }
