+++ src/main/java/it/polimi/ingsw/network/client/ServerActions.java
@@ -1,0 +1,4 @@
+package it.polimi.ingsw.network.client;
+
+public interface ServerActions {
+}
