--- src/main/java/it/polimi/ingsw/network/server/ServerActions.java
@@ -1,4 +1,0 @@
-package it.polimi.ingsw.network.server;
-
-public interface ServerActions {
-}
